refactor: update user context usage and improve code structure across components
